+++ AMD/Demo/amd_simple.c
@@ -15,7 +15,7 @@ int P [5] ;
 int main (void)
 {
     int k ;
-    (void) amd_order (n, Ap, Ai, P, (double *) NULL, (double *) NULL) ;
+    (void) amd_order (n, Ap, Ai, P, (float *) NULL, (float *) NULL) ;
     for (k = 0 ; k < n ; k++) printf ("P [%d] = %d\n", k, P [k]) ;
     return (0) ;
 }
